add new "jsonext" formatter to Console & File logger expose formatting codes into Formatted abstract logger to reduce redundancy.
